The one with the refactor
More refactoring and bug fixing because of refactoring. Pushing before
being distracted for a few days
