+++ _meta.lua
@@ -1,6 +1,5 @@
 local _ = require("gettext")
 return {
-  name = "airplanemode",
   fullname = _("AirPlaneMode"),
   description = _([[This plugin adds an AirPlaneMode.]]),
   version = "1.2.2",
